properties: add git.dirty.mark property
Motivation:

Easily include the dirty status in other maven properties, in the format
described by the 'dirty' configuration property.

Modification:

An additional property is exposed to maven: 'git.dirty.mark'.  If the
code-base has non-checked-in changes then the 'git.dirty.mark' property
value is the configured dirty mark, as used in 'git.commit.id.describe'.
If the code-base is clean then 'git.dirty.mark' expands to the empty
string.

Result:

Whether the code-base is clean or dirty is easily available in the
configured format.