Skip git invocation if properties are already computed
In a multi-modules project, one can set injectAllReactorProjects to
true to add all the properties to each reactor's project, but there's
no automated effort to skip git execution if properties are already set
unless runOnlyOnce is set to true.

But runOnce has two issues itself:
- it assumes that the top-level project is included in the reactor
- it skips the whole plugin execution

To address this, detect during plugin execution if properties from a
previous project in the reactor have been already computed and added
to the current project's context, and skip git execution if this is the
case.