Correct text bold formatting on license text
The bold text formatting is not working as there are is no whitespace
line between the `<img>` tag  and the line of text describing the
license under which this plugins is being released.

This commit adds a line of whitespace so that the bold text is
formatted correctly, and it also fixes a typo in one of the hyperlinks.