Introduce "offline" mode
Version 3.0.0 started to look at "ahead" and "behind" branches. To do
so, it executes a Fetch command to the repo. This is unfortunate
because it introduces quite a bit of a delay when building, especially
when the machine is offline (e.g. on a plane). Also, it sets off
various alerts because the previous version did not do any remote
access.

This PR introduces an "offline" mode (disables by default) that allows
turning off all commands that are trying to reach out to
remotes. Currently, that is only the fetch command which is executed
when trying to determine "ahead" and "behind". Any future operation
that is accessing remote repos should also check this flag.