Allow to override useNativeGit from command line
Right now, because JGit doesn't support Git worktrees, I need to override `useNativeGit`
in the `pom.xml` of different projects, and it's easy to make a mistake and commit the
changes.

The proposed PR should allow to override this property from command-line via
`maven.gitcommitid.nativegit`, without modification of the `pom.xml`.