[pom] Remove comment regarding now unused cobertura
While true, that isn't going to be fixed.  Jaccoc has been the defaco for years now.  The comment itself serves no purpose here as it's completely unlikely this will ever be switched back.  Cobertura doesn't support java 8.  We are at 13 now and 14 in 3 months.  The cobertura project is long EOL.  Also, project is nearly 5 years without a commit.