https://github.com/git-commit-id/git-commit-id-maven-plugin/issues/571: Remove location information when we publish under old releases
Welcome to a chicken or egg-problem.
In practice we want to relocate the plugin to a new location,
but if we do so without publishing the counterpart in the new location
maven fails with "Could not find artifact io.github.git-commit-id:git-commit-id-maven-plugin:jar:${version}".
If we publish a release under the new coordinates, maven silent ignores the plugin settings
when we happen to use the old coordinates in our config
(https://github.com/git-commit-id/git-commit-id-maven-plugin/issues/571).

As a result: The plugin is moving to a new location, but we can't encode it in the pom :-(

If anyone ever asks for a relocation pom again the answer is simply: no, thanks.